Integrated titlekey and padgen features.
