+++ source/decryptor/features.h
@@ -1,0 +1,6 @@
+#pragma once
+
+u32 NcchPadgen(void);
+u32 SdPadgen(void);
+u32 NandPadgen(void);
+u32 DecryptTitlekeys(void);
